Settings: Forcefully show more than 4 items in some menus
This is pretty much a fixup for 18f1e49926188541e75a40b1d5861a300b06c4fb
since some menus were still hiding settings.

A lot of the time, the settings I'm actually looking for are under
"advanced", which slows down my navigation to the target setting by
making me tap "Advanced" to expand it.

These hidden advanced sections are also bad for user discoverability
because it means that the user needs to expand all advanced sections in
order to discover all of the settings. It also makes the user hesitate
and scan through the non-advanced options first to determine whether the
option they're looking for is under advanced, since the advanced vs.
non-advanced grouping is often quite unintuitive. Realistically, with
discoverability and ease of access in mind, I can't think of a situation
in which the advanced sections will actually be helpful.

Change-Id: I6707682240112fab6e59e7b12512bea87b9b4655
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>